refactor(general): decompose bloated views into mixins and sub views
