+++ lib/features/community/discussion_detail/view/sub_view/discussion_header_title.dart
@@ -1,0 +1,32 @@
+part of '../discussion_detail_view.dart';
+
+final class _DiscussionHeaderTitle extends StatelessWidget {
+  const _DiscussionHeaderTitle({required this.args});
+
+  final DiscussionDetailArgs args;
+
+  @override
+  Widget build(BuildContext context) {
+    return Column(
+      mainAxisSize: MainAxisSize.min,
+      crossAxisAlignment: CrossAxisAlignment.start,
+      children: [
+        GeneralContentTitle(
+          value: args.discussion.title,
+          fontWeight: FontWeight.w700,
+          maxLine: AppConstants.kTwo,
+        ),
+        GeneralContentSmallTitle(
+          value: LocaleKeys.community_groupDetail_discussions_openedByGroup.tr(
+            namedArgs: {
+              'name': args.discussion.author.maskedDisplayName,
+              'group': args.group.name,
+            },
+          ),
+          color: AppColors.navy300,
+          maxLine: AppConstants.kOne,
+        ),
+      ],
+    );
+  }
+}
